+++ src/main/java/de/samply/query/Query.java
@@ -1,0 +1,7 @@
+package de.samply.query;
+
+public record Query(
+    String queryId,
+    String query,
+    QueryFormat queryFormat
+) {}