+++ src/main/java/de/samply/query/QueryFormat.java
@@ -1,0 +1,6 @@
+package de.samply.query;
+
+public enum QueryFormat {
+  FHIR,
+  CQL
+}