+++ src/main/java/de/samply/result/ResultFormat.java
@@ -1,0 +1,8 @@
+package de.samply.result;
+
+public enum ResultFormat {
+  CSV,
+  EXCEL,
+  JAVA_OBJECT,
+  FHIR
+}