+++ CHANGELOG.md
@@ -11,4 +11,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - LDM Client
 - Containers and Containers Templates
 - FHIR-Bundle to Containers Converter
-- FHIR Rev Includes in Containers Templates 
+- FHIR Rev Includes in Containers Templates
+- Executor in Fhir Store Client